requirements.txt
anyio==3.6.2
asgiref==3.5.2
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mpmath==1.2.1
passlib==1.7.4
Pillow==9.2.0
primePy==1.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
sympy==1.11.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.3
Werkzeug==2.2.2
zipp==3.9.0
